Move SwiftPhysicalLowering class to the common section
